Add option for running only the step 1 model
Enhanced the inference script and README by adding a `-step1` option to execute only the step 1 model, which outputs a single label for all vertebrae including the sacrum. Added conditional logic to skip step 2 processes if the `-step1` flag is used. Updated preview generation and iterative labeling steps accordingly. This feature helps users needing only step 1 results, saving processing time.

+++ scripts/inference_nnunet.sh
@@ -22,6 +22,7 @@ trap "echo Caught Keyboard Interrupt within script. Exiting now.; exit" INT
 
 INPUT_FOLDER="$1"
 OUTPUT_FOLDER="$2"
+if [[ $3 == -step1 ]]; then STEP1=1; else STEP1=0; fi
 
 # ensure the custom nnUNetTrainer is defined in the nnUNet library and add it if it is not
 source "$TOTALSPINESEG"/scripts/add_nnunet_trainer.sh
@@ -87,7 +88,7 @@ if [ ! -d "$nnUNet_results"/Dataset${step2_dataset}_* ]; then
     nnUNetv2_install_pretrained_model_from_zip "$nnUNet_exports"/Dataset${step2_dataset}_*.zip
 fi
 
-#Convert 4D images to 3D
+# Make output dir with copy of the input images convert 4D images to 3D
 totalspineseg_average4d -i "${INPUT_FOLDER}" -o "${OUTPUT_FOLDER}"/input --image-suffix "" --output-image-suffix "" -r
 
 # Add _0000 to inputs if not exists to run nnunet
@@ -96,45 +97,61 @@ for f in "${OUTPUT_FOLDER}"/input/*.nii.gz; do if [[ "$f" != *_0000.nii.gz ]]; t
 #Transform images to canonical space
 totalspineseg_reorient_canonical -i "${OUTPUT_FOLDER}"/input -o "${OUTPUT_FOLDER}"/input -r
 
-# Make output dir with copy of the input images resampled to 1x1x1mm
+# resampled to 1x1x1mm
 totalspineseg_resample -i "${OUTPUT_FOLDER}"/input -o "${OUTPUT_FOLDER}"/input --image-suffix "" --output-image-suffix "" -r
+
+# Generate preview images
+totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -o "${OUTPUT_FOLDER}"/preview --output-suffix _input -r
 
 # Run step 1 model
 # Check if the final checkpoint exists, if not use the latest checkpoint
 if [ -f "$nnUNet_results"/Dataset${step1_dataset}_*/${nnUNetTrainer}__${nnUNetPlans}__${configuration}/fold_${FOLD}/checkpoint_final.pth ]; then CHECKPOINT=checkpoint_final.pth; else CHECKPOINT=checkpoint_latest.pth; fi
 nnUNetv2_predict -d $step1_dataset -i "${OUTPUT_FOLDER}"/input -o "${OUTPUT_FOLDER}"/step1 -f $FOLD -c $configuration -p $nnUNetPlans -tr $nnUNetTrainer -npp $JOBS -nps $JOBS -chk $CHECKPOINT -device $DEVICE
 
+# Generate preview images for step 1
+totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/step1 -o "${OUTPUT_FOLDER}"/preview --output-suffix _step1 -r
+
 # Extract the largest connected component of step 1 segmentation after binarization and dilation into the input folder of step 2 model
-totalspineseg_largest_component -s "${OUTPUT_FOLDER}"/step1 -o "${OUTPUT_FOLDER}"/step2_input --output-seg-suffix _0001 --binarize --dilate 5 -r
+totalspineseg_largest_component -s "${OUTPUT_FOLDER}"/step1 -o "${OUTPUT_FOLDER}"/output --binarize --dilate 5 -r
 
-# Copy the original image into the input folder of step 2 model to use as the 1'st channel
-totalspineseg_cpdir "${OUTPUT_FOLDER}"/input "${OUTPUT_FOLDER}"/step2_input -p "*_0000.nii.gz" -r
+# We use an iterative algorithm to label IVDs with the definite labels, based on the C2-C3, C7-T1 and L5-S1 IVD labels output by step 1 model.
+totalspineseg_iterative_label -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/output --disc-labels 1 2 3 4 5 --init-disc 2:224 5:202 3:219 4:207 --output-disc-step -1 --map-input 6:92 7:201 8:201 9:200 -r
 
-# Crop the images to the bounding box of the non-zero part of the step 1 segmentation (with 10 voxels margin), this will also delete images without segmentation
-totalspineseg_crop_image2seg -i "${OUTPUT_FOLDER}"/step2_input -s "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2_input --seg-suffix _0001 -m 10 -r
+if [ $STEP1 -eq 0 ]; then
 
-# Transform step 1 segmentation to the cropped images space
-totalspineseg_transform_seg2image -i "${OUTPUT_FOLDER}"/step2_input -s "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2_input --seg-suffix _0001 --output-seg-suffix _0001 -r
+    # Copy the original image into the input folder of step 2 model to use as the 1'st channel
+    totalspineseg_cpdir "${OUTPUT_FOLDER}"/input "${OUTPUT_FOLDER}"/step2_input -p "*_0000.nii.gz" -r
 
-# Distinguished odd IVDs based on the C2-C3, C7-T1 and L5-S1 IVD labels output by step 1 model:
-# First we will use an iterative algorithm to label IVDs with the definite labels
-totalspineseg_iterative_label -s "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2_input --seg-suffix _0001 --output-seg-suffix _0001 --disc-labels 1 2 3 4 5 --init-disc 2:224 5:202 3:219 4:207 --output-disc-step -1 -r
-# Then, we map the IVDs labels to the odd IVDs to use as the 2'nd channel, this will also delete labels without odd IVDs
-totalspineseg_map_labels -s "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2_input -m "$resources"/labels_maps/nnunet_step2_input.json --seg-suffix _0001 --output-seg-suffix _0001 -r
+    # Crop the images to the bounding box of the non-zero part of the step 1 segmentation (with 10 voxels margin), this will also delete images without segmentation
+    totalspineseg_crop_image2seg -i "${OUTPUT_FOLDER}"/step2_input -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/step2_input -m 10 -r
 
-# Remove images without the 2'nd channel
-for f in "${OUTPUT_FOLDER}"/step2_input/*_0000.nii.gz; do if [ ! -f "${f/_0000.nii.gz/_0001.nii.gz}" ]; then rm $f; fi; done
+    # Transform step 1 segmentation to the cropped images space
+    totalspineseg_transform_seg2image -i "${OUTPUT_FOLDER}"/step2_input -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/step2_input --output-seg-suffix _0001 -r
 
-# Run step 2 model
-# Check if the final checkpoint exists, if not use the latest checkpoint
-if [ -f "$nnUNet_results"/Dataset${step1_dataset}_*/${nnUNetTrainer}__${nnUNetPlans}__${configuration}/fold_${FOLD}/checkpoint_final.pth ]; then CHECKPOINT=checkpoint_final.pth; else CHECKPOINT=checkpoint_latest.pth; fi
-nnUNetv2_predict -d $step2_dataset -i "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2 -f $FOLD -c $configuration -p $nnUNetPlans -tr $nnUNetTrainer -npp $JOBS -nps $JOBS -chk $CHECKPOINT -device $DEVICE
+    # Now, we map the IVDs labels from the step1 model output to the odd IVDs to use as the 2'nd channel, this will also delete labels without odd IVDs
+    totalspineseg_map_labels -s "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2_input -m "$resources"/labels_maps/nnunet_step2_input.json --seg-suffix _0001 --output-seg-suffix _0001 -r
 
-# Extract the largest connected component of step 2 segmentation after binarization and dilation into the output folder
-totalspineseg_largest_component -s "${OUTPUT_FOLDER}"/step2 -o "${OUTPUT_FOLDER}"/output --binarize --dilate 5 -r
+    # Remove images without the 2'nd channel
+    for f in "${OUTPUT_FOLDER}"/step2_input/*_0000.nii.gz; do if [ ! -f "${f/_0000.nii.gz/_0001.nii.gz}" ]; then rm $f; fi; done
 
-# Use an iterative algorithm to to assign an individual label value to each vertebrae and IVD in the final segmentation mask.
-totalspineseg_iterative_label -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/output --disc-labels 2 3 4 5 6 7 --vertebrea-labels 9 10 11 12 13 14 --init-disc 4:224 7:202 5:219 6:207 --init-vertebrae 11:40 14:17 12:34 13:23 --step-diff-label --step-diff-disc --output-disc-step -1 --output-vertebrea-step -1 --map-output 40:92 --map-input 14:92 16:201 17:200 -r
+    # Generate preview images for step 2 input
+    totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/preview --seg-suffix _0001 --output-suffix _step2_input -r
+
+    # Run step 2 model
+    # Check if the final checkpoint exists, if not use the latest checkpoint
+    if [ -f "$nnUNet_results"/Dataset${step1_dataset}_*/${nnUNetTrainer}__${nnUNetPlans}__${configuration}/fold_${FOLD}/checkpoint_final.pth ]; then CHECKPOINT=checkpoint_final.pth; else CHECKPOINT=checkpoint_latest.pth; fi
+    nnUNetv2_predict -d $step2_dataset -i "${OUTPUT_FOLDER}"/step2_input -o "${OUTPUT_FOLDER}"/step2 -f $FOLD -c $configuration -p $nnUNetPlans -tr $nnUNetTrainer -npp $JOBS -nps $JOBS -chk $CHECKPOINT -device $DEVICE
+
+    # Generate preview images for step 2
+    totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/step2 -o "${OUTPUT_FOLDER}"/preview --output-suffix _step2 -r
+
+    # Extract the largest connected component of step 2 segmentation after binarization and dilation into the output folder
+    totalspineseg_largest_component -s "${OUTPUT_FOLDER}"/step2 -o "${OUTPUT_FOLDER}"/output --binarize --dilate 5 -r
+
+    # Use an iterative algorithm to to assign an individual label value to each vertebrae and IVD in the final segmentation mask.
+    totalspineseg_iterative_label -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/output --disc-labels 2 3 4 5 6 7 --vertebrea-labels 9 10 11 12 13 14 --init-disc 4:224 7:202 5:219 6:207 --init-vertebrae 11:40 14:17 12:34 13:23 --step-diff-label --step-diff-disc --output-disc-step -1 --output-vertebrea-step -1 --map-output 40:92 --map-input 14:92 16:201 17:200 -r
+
+fi
 
 # Fill spinal cancal label to include all non cord spinal canal, this will put the spinal canal label in all the voxels (labeled as a backgroupn) between the spinal canal and the spinal cord.
 totalspineseg_fill_canal -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/output --canal-label 201 --cord-label 200 --largest-canal --largest-cord -r
@@ -142,8 +159,5 @@ totalspineseg_fill_canal -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/outp
 # Transform labels to input images space
 totalspineseg_transform_seg2image -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/output -r
 
-# Generate preview images
-totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -o "${OUTPUT_FOLDER}"/preview --output-suffix _input -r
-totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/step1 -o "${OUTPUT_FOLDER}"/preview --output-suffix _step1 -r
-totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/step2 -o "${OUTPUT_FOLDER}"/preview --output-suffix _step2 -r
+# Generate preview images for the final output
 totalspineseg_preview_jpg -i "${OUTPUT_FOLDER}"/input -s "${OUTPUT_FOLDER}"/output -o "${OUTPUT_FOLDER}"/preview --output-suffix _output -r
